check transfer history about rewarding in Post page
